8.30额。。移除了悬浮窗。。

+++ app/src/main/java/com/example/rungame10/biyue/View/SwitchDialog.java
@@ -4,22 +4,23 @@ import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.SharedPreferences;
+import android.content.res.Configuration;
 import android.os.Bundle;
 import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
-import android.widget.PopupWindow;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.example.rungame10.biyue.Presenter.FloatActionController;
+import com.example.rungame10.biyue.Common.Config;
+import com.example.rungame10.biyue.Intf.ResultCode;
 import com.example.rungame10.biyue.Util.MResource;
 
 public class SwitchDialog extends AlertDialog {
     private Context context;
     private TextView nameText,logoutBtn;
+    private boolean logoutFlag = false;
 
     public SwitchDialog(Context context){
         super(context, MResource.getIdByName(context, "style", "by_Dialog"));
@@ -49,13 +50,24 @@ public class SwitchDialog extends AlertDialog {
         logoutBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                SharedPreferences.Editor editor = sharedPreferences.edit();
-                editor.clear();
-                editor.apply();
-                FloatActionController.isLogined = false;
-                FloatActionController.getInstance().stopServer(context);
-                SwitchDialog.this.cancel();
-                Toast.makeText(context,"用户退出登录成功，请重新登录",Toast.LENGTH_SHORT).show();
+                logoutFlag = true;
+                cancel();
+            }
+        });
+
+        setOnCancelListener(new OnCancelListener() {
+            @Override
+            public void onCancel(DialogInterface dialogInterface) {
+                if (logoutFlag){
+                    SharedPreferences.Editor editor = sharedPreferences.edit();
+                    editor.clear();
+                    editor.apply();
+                    Config.isLogined = false;
+                    Config.logoutCallBack.onResponse(ResultCode.LOGOUT_SUCCESS);
+                }else {
+                    Config.isLogined = true;
+                    Config.loginCallBack.onResponse(ResultCode.LOGIN_SUCCESS);
+                }
             }
         });
 
@@ -65,6 +77,21 @@ public class SwitchDialog extends AlertDialog {
         DisplayMetrics d = context.getResources().getDisplayMetrics();     //获取屏幕宽高
         lp.width = (int) (d.widthPixels*0.8);
         lp.height = (int) (d.heightPixels*0.3);
+
+        //判断当前是否横屏
+        Configuration configuration = context.getResources().getConfiguration();
+        int ori = configuration.orientation;
+        if(ori == Configuration.ORIENTATION_LANDSCAPE){
+            //横屏
+            lp.width = (int) (d.widthPixels*0.5);
+            lp.height = (int) (d.heightPixels*0.7);
+
+        }else if(ori == Configuration.ORIENTATION_PORTRAIT){
+            //竖屏
+            lp.width = (int) (d.widthPixels*0.8);
+            lp.height = (int) (d.heightPixels*0.3);
+        }
+
         dialogWindow.setAttributes(lp);
     }
 }
